Fix generating MsgSendSmallInt.bc with clang trunk.  Chris Lattner 'helpfully' removed the overflow handler support from clang without bothering to tell anyone, so -ftrapv in clang now does the same (rather useless) thing that it does in GCC.  I've now added -ftrapv-handler= which lets us restore the old behaviour.
Note: Now that handlers can be generated per compilation unit, we have some interesting possibilities.  Currently, we overflow to a BigInt, however it would be trivial to overflow to a BoxedFloat (just recompile MsgSendSmallInt with a different -ftrapv-handler= which names a function where BigInt is replaced with BoxedFloat in the handler).  This would permit us to implement JavaScript-compatible number semantics without any extra effort.

I'm not really motivated to do this for real code, because JavaScript's overflow semantics are braindead, but it would make a nice demo...



git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Languages@6344 7b3f36aa-e6db-49c1-8b4c-66eac1997e64